Dynamically determine unpack directory for Yarn
